Move ImageSerializer to People app
This keeps us in line with the pattern started in the parties app
