+++ ynr/apps/people/serializers.py
@@ -1,0 +1,27 @@
+from rest_framework import serializers
+
+from people.models import PersonImage
+
+
+class ImageSerializer(serializers.HyperlinkedModelSerializer):
+    class Meta:
+        model = PersonImage
+        fields = (
+            "id",
+            "url",
+            "source",
+            "is_primary",
+            "md5sum",
+            "copyright",
+            "uploading_user",
+            "user_notes",
+            "user_copyright",
+            "notes",
+            "image_url",
+        )
+
+    image_url = serializers.SerializerMethodField()
+    uploading_user = serializers.ReadOnlyField(source="uploading_user.username")
+
+    def get_image_url(self, instance):
+        return instance.image.url
